Add support for deleting surveys
