+++ src/css/sass/components/surveys.scss
@@ -257,3 +257,17 @@
   }
 }
 
+
+/* Settings page ------------------------------- */
+.survey-settings {
+  padding-bottom: 15px;
+}
+
+#delete-view-container {
+  border-top: 6px solid #fff;
+  margin-top: 1em;
+}
+
+#delete-survey-form {
+  display: none;
+}
