+++ src/css/sass/shared/buttons.scss
@@ -85,6 +85,7 @@ button.danger {
 
 button.danger:hover {
   background: #f36c3c;
+  border-color: #f36c3c;
 }
 
 button.disabled {
