+++ src/js/templates/surveys/settings.html
@@ -32,5 +32,6 @@
   </div>
   <div class="b">
     <div id="sharing-view-container"></div>
+    <div id="delete-view-container"></div>
   </div>
 </div>
