+++ src/js/views/surveys/settings.js
@@ -1,230 +1,85 @@
 /*jslint nomen: true */
-/*globals define, cartodb, Rickshaw: true */
+/*globals define: true */
 
-// settings-survey
-//
-define(function (require) {
+define(function(require, exports, module) {
   'use strict';
 
+  // Libs
   var $ = require('jquery');
   var _ = require('lib/lodash');
   var Backbone = require('backbone');
 
-  // LocalData
-  var settings = require('settings');
-
-  // Models
-  var Responses = require('models/responses');
+  // Views
+  var DeleteView = require('views/surveys/delete');
+  var MapDrawView = require('views/maps/draw');
+  var UsersView = require('views/surveys/users');
 
   // Templates
-  var template = require('text!templates/projects/surveys/settings-survey.html');
-  var loadingTemplate = require('text!templates/filters/loading.html');
+  var template = require('text!templates/surveys/settings.html');
 
-  var ANSWER = 'response';
-  var NOANSWER = 'no response';
 
-  /**
-   * Intended for shorter lists of responses (arbitrarily <25)
-   * Doesn't include pagination, which isn't relevant in this case.
-   * See responses/responses/ListView for a heavyweight implementation.
-   */
-  var FilterView = Backbone.View.extend({
-    className: 'settings',
-    filters: {},
-
+  var SettingsView = Backbone.View.extend({
+    el: '#settings-view-container',
     template: _.template(template),
-    loadingTemplate: _.template(loadingTemplate),
 
     events: {
-      "click .select": "showOptions",
-      "click .question-title": "selectQuestion",
-      "click .answer": "selectAnswer",
-      "click .clear": "reset",
-      "click .close-settings": "close"
+      'click .save': 'save'
     },
 
     initialize: function(options) {
-      _.bindAll(this, 'render', 'reset');
+      _.bindAll(this, 'render', 'save', 'success', 'error');
 
       this.survey = options.survey;
-      this.forms = options.forms;
-      // this.mapView = options.map;
 
-      this.stats = options.stats;
-      this.stats.on('change', this.render);
-
-      this.$el.html(this.loadingTemplate({}));
+      this.survey.on('change', this.render);
     },
 
-    close: function() {
-      console.log("Closing settings");
-      this.$el.hide();
+    error: function(model, xhr, options) {
+      $("#survey-settings-form .error").fadeIn().css("display","inline-block").delay(2000).fadeOut();
     },
 
-    prepQuestions: function(questions, stats) {
-      // Match the question names and answer values from the form with stats and colors.
-      _.each(_.keys(questions), function (question) {
-        var questionStats = stats.get(question);
-        var type = questions[question].type;
-        if (type === 'text') {
-          var total = _.reduce(questionStats, function (sum, count) {
-            return sum + count;
-          }, 0);
+    success: function() {
+      $("#survey-settings-form .saved").fadeIn().css("display","inline-block").delay(2000).fadeOut();
+    },
 
-          var noResponseCount;
-          if (questionStats) {
-            noResponseCount = questionStats[NOANSWER];
-          }
-          if (noResponseCount === undefined) {
-            noResponseCount = 0;
-          }
+    save: function(event) {
+      event.preventDefault();
 
-          questions[question].answers = [{
-            text: ANSWER,
-            value: ANSWER,
-            count: total - noResponseCount,
-            color: settings.colorRange[1]
-          }, {
-            text: NOANSWER,
-            value: NOANSWER,
-            count: noResponseCount,
-            color: settings.colorRange[0]
-          }];
-        } else if (type === 'file') {
-          // TODO: We need to see photo upload numbers in the stats (or
-          // somewhere) to report them in the UI.
+      // Get the fields from the form
+      var form = $(event.target).parent().serializeArray();
 
-          questions[question].answers = [{
-            text: ANSWER,
-            value: ANSWER,
-            count: '',
-            color: settings.colorRange[1]
-          }, {
-            text: NOANSWER,
-            value: NOANSWER,
-            count: '',
-            color: settings.colorRange[0]
-          }];
-        } else {
-          var answers = questions[question].answers;
-          _.each(answers, function (answer, index) {
-            if (!questionStats) {
-              answer.count = 0;
-            } else {
-              // Get the count from the stats object.
-              answer.count = questionStats[answer.value];
-              if (answer.count === undefined) {
-                answer.count = 0;
-              }
-            }
+      // Convert the fields into a format that can be saved
+      var fields = _.reduce(form, function(memo, field) {
+        memo[field.name] = field.value;
+        return memo;
+      }, {});
 
-            // Get the color.
-            // The last "answer" is the no-response placeholder, which gets the
-            // zero-index color.
-            answer.color = settings.colorRange[(index + 1) % answers.length];
-          });
-        }
+      this.survey.set(fields);
+      this.survey.attributes.zones = this.mapDrawView.getZones();
+      this.survey.save({}, {
+        success: this.success,
+        error: this.error
       });
-
-      return questions;
     },
 
     render: function() {
-      console.log('Rendering the filters');
-
-      var questions = this.forms.getFlattenedForm();
-      var stats = this.stats;
-      questions = this.prepQuestions(questions, stats);
-
       var context = {
-        questions: questions,
-        mapping: this.forms.map()
+        survey: this.survey.toJSON()
       };
+
       this.$el.html(this.template(context));
-      return this.$el;
-    },
 
-    showOptions: function(event) {
-      event.preventDefault();
-      $('.filters .options').show();
-    },
-
-    /**
-     * Reset any filters
-     */
-    reset: function(event) {
-      if(event) {
-        event.preventDefault();
-      }
-
-      this.filters = {};
-      this.trigger('filterReset');
-
-      $('.filters .clear').slideUp();
-      $('.filters .options .answers').slideUp();
-      $('.filters .answer').removeClass('active');
-    },
-
-    selectQuestion: function(event) {
-      // Clear out any filters
-      if(this.filters.answer) {
-        this.reset();
-      }
-
-      // Set up handy shortcuts
-      var $question = $(event.target).parent();
-      var question = $question.attr('data-question');
-      this.filters.question = question;
-
-      // Show the sub-answers
-      $('.filters .options .answers').slideUp();
-      $question.find('.answers').slideDown();
-      $('.filters .clear').slideDown();
-      $question.find('.toggle').slideUp();
-
-      this.trigger('filterSet', this.filters);
-
-      // XXX TODO
-      // Mark this question as selected
-    },
-
-    /**
-     * Show only responses with a specific answer
-     */
-    selectAnswer: function(event) {
-      var $answer = $(event.target);
-      this.filters.answer = $answer.attr('data-answer');
-
-      // Make sure we have the right question selected
-      this.filters.question = $answer.attr('data-question');
-      var $question = $('div[data-question=' + this.filters.question + ']');
-
-      if(!this.filters.answer) {
-        $answer = $answer.parent();
-        this.filters.answer = $answer.attr('data-answer');
-      }
-
-      this.trigger('filterSet', this.filters);
-
-      $('.filters .answer').removeClass('active');
-      $answer.addClass('active');
-    },
-
-    /**
-     * Associate a unqie color with each answer in a list
-     */
-    colors: function(keys) {
-      var answers = {};
-      _.each(keys, function(key, index) {
-        answers[key] = {
-          color: settings.colorRange[index + 1]
-        };
+      this.mapDrawView = new MapDrawView({
+        survey: this.survey
       });
-      return answers;
+      this.sharingView = new UsersView({
+        survey: this.survey
+      });
+      this.deleteView = new DeleteView({
+        survey: this.survey
+      });
     }
-
   });
 
-  return FilterView;
-
+  return SettingsView;
 });
